feat: update AVA configuration for test file matching and add CI workflow for automated testing
