+++ .github/workflows/ci.yml
@@ -1,0 +1,17 @@
+on:
+  release:
+    types:
+      - published
+
+jobs:
+  publish:
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v4
+      - uses: actions/setup-node@v4
+        with:
+          node-version: 22.9
+      - run: npm ci
+
+      - name: Run Tests
+        run: npm run test
